Rename validators and migrate some syntactic validation unit tests.
This commit will also skip `only_one_storage` validation if `partial` is `True`.
