+++ cli/src/pcluster/validators/ec2_validators.py
